Move types tests
