+++ packages/types/.mocharc.json
@@ -1,6 +1,7 @@
 {
-  "spec": "test/**/*.test.ts",
+  "spec": "{src,test}/**/*.test.ts",
   "require": "esbuild-register",
   "watchExtensions": "ts",
-  "extension": "ts"
+  "extension": "ts",
+  "reporterOption": "maxDiffSize=0"
 }
